updating build.properties and build.xml to make it easier to build contrib projects.  Moved the configuration for contrib projects into build.properties so there is no more editing of build.xml required to build various contrib projects.
In order to add new contrib projects to build, just edit the src.contrib.projects property in build.properties with the path to the new project to be built.  Make sure the project builds first, though.  Some stuff under the "contrib" directory won't build which is why we specifically point to those that do build.

I also modified the names of targets which shouldn't be directly called from the command line by prepending a dash to each of them.  This makes the build more fail-safe.  Also added CVS version information to the top of both files to make it easier to distinguish what the version of the file is and when it was last edited.

Jake