Keep using the existing journal file if creating a new one fails.
Eg. If the process in which Prevayler is executing exceeds its open
files limit it will be unable to open a new journal but can keep writing
to the existing one without any problems.

This is useful for systems that would prefer to handle this scenario
by trying to keep limping rather than failing hard.