Make E103Test more reliable.
It was occasionally failing if replaying the journal took
less than 1 millisecond, because then even the example
incorrect transaction would get the right answer. Sleeping
for 1ms before replaying the journal ensures that the
timestamps are different.

(Technically, the clock doesn't necessarily have millisecond
accuracy, so this still isn't absolutely deterministic, but is
likely to be good enough on modern machines.)

The test was also creating temporary directories with
millisecond timestamps in the name, and not cleaning them up,
which could potentially be problematic. Now they use a single
temporary directory and ensure that it is empty before running.

Fixes #30